Move shared pages into shared lib module
