--- app/controllers/hello_world_controller.rb
@@ -1,7 +1,0 @@
-class HelloWorldController < ApplicationController
-  layout "hello_world"
-
-  def index
-    @hello_world_props = { name: "Stranger" }
-  end
-end
