--- client/app/bundles/AdminView/components/SettingsPage/FeedbackContainer.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const FeedbackContainer = () =>
-  <div className="settings-pane">
-    <h1>Submit Feedback</h1>
-    <p>
-      <em>
-        This feature isn{"'"}t quite ready yet. Come back later!
-      </em>
-    </p>
-  </div>;
-
-export default FeedbackContainer;
