--- client/app/bundles/AdminView/components/SettingsPage/TeamSettingsContainer.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const TeamSettingsContainer = () =>
-  <div className="settings-pane">
-    <h1>Team Settings</h1>
-    <p>
-      <em>
-        This feature isn{"'"}t quite ready yet. Come back later!
-      </em>
-    </p>
-  </div>;
-
-export default TeamSettingsContainer;
