--- client/app/bundles/HackerView/components/SettingsPage/NotificationSettingsContainer.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const NotificationsSettingsContainer = () =>
-  <div className="settings-pane">
-    <h1>Notifications Settings</h1>
-    <p>
-      <em>
-        This feature isn{"'"}t quite ready yet. Come back later!
-      </em>
-    </p>
-  </div>;
-
-export default NotificationsSettingsContainer;
