--- client/app/bundles/HackerView/components/SettingsPage/WithdrawContainer.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const WithdrawContainer = () =>
-  <div className="settings-pane">
-    <h1>Withdraw Your Place</h1>
-    <p>
-      <em>
-        This feature isn{"'"}t quite ready yet. Come back later!
-      </em>
-    </p>
-  </div>;
-
-export default WithdrawContainer;
