--- client/app/bundles/HelloWorld/actions/helloWorldActionCreators.jsx
@@ -1,8 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-
-import { HELLO_WORLD_NAME_UPDATE } from "../constants/helloWorldConstants";
-
-export const updateName = text => ({
-  type: HELLO_WORLD_NAME_UPDATE,
-  text,
-});
