--- client/app/bundles/HelloWorld/constants/helloWorldConstants.jsx
@@ -1,3 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-
-export const HELLO_WORLD_NAME_UPDATE = "HELLO_WORLD_NAME_UPDATE";
