--- client/app/bundles/HelloWorld/startup/registration.jsx
@@ -1,8 +1,0 @@
-import ReactOnRails from "react-on-rails";
-
-import HelloWorldApp from "./HelloWorldApp";
-
-// This is how react_on_rails can see the HelloWorld in the browser.
-ReactOnRails.register({
-  HelloWorldApp,
-});
