--- client/app/bundles/HelloWorld/store/helloWorldStore.jsx
@@ -1,6 +1,0 @@
-import { createStore } from "redux";
-import helloWorldReducer from "../reducers/helloWorldReducer";
-
-const configureStore = railsProps => createStore(helloWorldReducer, railsProps);
-
-export default configureStore;
